requirements.txt
mkdocs
mkdocs-material
mkdocs-glightbox
mkdocs-macros-plugin
python-slugify
regex